32-bit necrocomputing.

+++ tests/hmproto.h
@@ -4,10 +4,10 @@
     void *x##_new(void);\
     void x##_delete(void *c);\
     \
-    int x##_insert(void *c, uint64_t key, void *value, int update);\
-    void *x##_remove(void *c, uint64_t key);\
-    void *x##_get(void *c, uint64_t key);\
-    void *x##_find_le(void *c, uint64_t key);
+    int x##_insert(void *c, uintptr_t key, void *value, int update);\
+    void *x##_remove(void *c, uintptr_t key);\
+    void *x##_get(void *c, uintptr_t key);\
+    void *x##_find_le(void *c, uintptr_t key);
 
 //HM_PROTOS(critbit)
 //HM_PROTOS(tcradix)
@@ -16,10 +16,10 @@ HM_PROTOS(critnib)
 
 void *(*hm_new)(void);
 void (*hm_delete)(void *c);
-int (*hm_insert)(void *c, uint64_t key, void *value, int update);
-void *(*hm_remove)(void *c, uint64_t key);
-void *(*hm_get)(void *c, uint64_t key);
-void *(*hm_find_le)(void *c, uint64_t key);
+int (*hm_insert)(void *c, uintptr_t key, void *value, int update);
+void *(*hm_remove)(void *c, uintptr_t key);
+void *(*hm_get)(void *c, uintptr_t key);
+void *(*hm_find_le)(void *c, uintptr_t key);
 const char *hm_name;
 int hm_immutable;
 
@@ -39,10 +39,10 @@ struct hm
 {
     void *(*hm_new)(void);
     void (*hm_delete)(void *c);
-    int (*hm_insert)(void *c, uint64_t key, void *value, int update);
-    void *(*hm_remove)(void *c, uint64_t key);
-    void *(*hm_get)(void *c, uint64_t key);
-    void *(*hm_find_le)(void *c, uint64_t key);
+    int (*hm_insert)(void *c, uintptr_t key, void *value, int update);
+    void *(*hm_remove)(void *c, uintptr_t key);
+    void *(*hm_get)(void *c, uintptr_t key);
+    void *(*hm_find_le)(void *c, uintptr_t key);
     const char *hm_name;
     int hm_immutable;
 } hms[1];
